Comparison graph working in all show pages
