+++ app/javascript/components/comparison_chart.js
@@ -1,52 +1,3 @@
-const userC = (userStats) => {
-  Chart.defaults.global.defaultFontColor = 'rgba(232, 205, 162, 1)';
-  Chart.defaults.global.defaultFontSize = 14;
-  let ctx = document.querySelector('#userChart');
-  new Chart(ctx, {
-    type: 'radar',
-    data: {
-      labels: ['Close Combat', 'Long Range Weapons', 'Mid Range Weapons', 'Explosives', 'Infiltration', 'Hacking', 'Lockpicking', 'Seduction'],
-      
-      datasets: [{
-        label: 'Skillset',
-        data: [
-          userStats.close_combat,
-          userStats.long_range_weapons,
-          userStats.mid_range_weapons,
-          userStats.explosives,
-          userStats.infiltration,
-          userStats.hacking,
-          userStats.lockpicking,
-          userStats.seduction
-        ],
-        backgroundColor: ['rgba(219, 159, 61, 0.7)'],
-        borderColor: ['rgba(219, 159, 61, 1)'],
-        borderWidth: 1
-      }],
-    },
-    options: {
-      scale: {
-        angleLines: {
-          color: 'rgba(232, 205, 162, 0.4)'
-        },
-        gridLines: {
-          color: 'rgba(232, 205, 162, 0.4)'
-        },
-        ticks: {
-          beginAtZero: true,
-          max: 100,
-          backdropColor: ['rgba(211, 171, 106, 0)'],
-          stepSize: 25
-        },
-        pointLabels: {
-          fontSize: 14
-        }
-      },
-      legend: { display: false },
-    }
-  });
-};
-
 const crimeC = (userStats, crimeStats) => {
   Chart.defaults.global.defaultFontColor = 'rgba(232, 205, 162, 1)';
   Chart.defaults.global.defaultFontSize = 14;
@@ -112,15 +63,10 @@ const crimeC = (userStats, crimeStats) => {
   });
 };
 
-const userChart = () => {
-  $.get('/current_user_json', function (userStats) {
-    console.log('testing current_user_json response', userStats.close_combat);
-    userC(userStats);
-  });
-}
+const crimeChart = (paramsId) => {
+  console.log('params_id:', paramsId);
 
-const crimeChart = () => {
-  $.get('/crimes/85/crime_json', function (crimeStats) {
+  $.get(`/crimes/${paramsId}/crime_json`, function (crimeStats) {
     $.get('/current_user_json', function (userStats) {
       console.log('user in join chart', userStats.hacking);
       console.log('crime in join chart', crimeStats.hacking);
@@ -129,4 +75,4 @@ const crimeChart = () => {
   });
 }
 
-export { userChart, crimeChart };
+export { crimeChart };
