+++ app/javascript/packs/application.js
@@ -26,14 +26,17 @@ require("channels")
 import "bootstrap";
 
 // Internal imports, e.g:
-import { crimeChart, userChart } from '../components/radar_chart';
+import { userChart } from '../components/user_chart';
+import { crimeChart } from '../components/comparison_chart';
 import { toggleDashboard, toggleNewEventPage } from '../components/show_page';
 import { toggleTabs, activateTab } from '../components/toggle_tabs';
 import { toggleCrimeFormCard, toggleSkillsetFormCard } from "../components/form_progression";
 
+let paramsId = document.getElementById("paramsId").innerHTML;
+
 document.addEventListener('turbolinks:load', () => {
   userChart();
-  crimeChart();
+  crimeChart(paramsId);
   toggleTabs();
   activateTab();
   toggleNewEventPage();
