+++ app/views/crimes/show.html.erb
@@ -74,6 +74,8 @@
   </div>
 </div>
 
+<%# This is used to pass the crime's id to the chart %>
+<p id='paramsId' style="visibility: hidden;"><%= @crime.id %></p>
 
 
 <!-- only bosses and right hands can see these two. Need to add logic for user -->
